initial framework branch, moving click_counter to its own file.
